Include QReport and friends in MEData.
